+++ bullet-2.81-rev2613/Demos/WWD/MTree.h
@@ -12,7 +12,7 @@ public:
 		NNPart = new std::vector<NNLayerNode*>();
 	}
 	~MTree(void){
-		for(int i=0;i<NNPart->size();i++){
+		for(int i=0;i<(int)NNPart->size();i++){
 			delete NNPart->at(i);
 		}
 		delete bodyPart;
